added content to fun facts page
